// src/components/FilterRow/type.ts
export enum FilterType {
  NONE = 'NONE',
  Normal = 'Normal',
  Emboss = 'Emboss',
  Earlybird = 'Earlybird',
  FuzzyGlass = 'FuzzyGlass',
  Tint = 'Tint',
  Warm = 'Warm',
  Cool = 'Cool',
  Saturate = 'Saturate',
  Invert = 'Invert',
  Nightvision = 'Nightvision',
  Technicolor = 'Technicolor',
  Polaroid = 'Polaroid',
  ToBGR = 'ToBGR',
  Kodachrome = 'Kodachrome',
  Browni = 'Browni',
  Vintage = 'Vintage',
  Lsd = 'Lsd',
  Protanomaly = 'Protanomaly',
  Deuteranomaly = 'Deuteranomaly',
  Tritanomaly = 'Tritanomaly',
  Protanopia = 'Protanopia',
  Achromatopsia = 'Achromatopsia',
  Achromatomaly = 'Achromatomaly',
}